🔨 Rename Identity namespace
