+++ src/SAF.DAS.ApprenticeCommitments.Web.UnitTests/Features/CannotConfirmSteps.cs
@@ -3,7 +3,7 @@ using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using FluentAssertions;
 using SFA.DAS.ApprenticeCommitments.Web.Pages.Apprenticeships;
-using SFA.DAS.ApprenticeCommitments.Web.Pages.IdentityHashing;
+using SFA.DAS.ApprenticeCommitments.Web.Identity;
 using TechTalk.SpecFlow;
 
 namespace SFA.DAS.ApprenticeCommitments.Web.UnitTests.Features
