+++ src/SAF.DAS.ApprenticeCommitments.Web.UnitTests/Features/MyApprenticeshipsSteps.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using SFA.DAS.ApprenticeCommitments.Web.Pages.Apprenticeships;
-using SFA.DAS.ApprenticeCommitments.Web.Pages.IdentityHashing;
+using SFA.DAS.ApprenticeCommitments.Web.Identity;
 using System.Net;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
