+++ src/SAF.DAS.ApprenticeCommitments.Web.UnitTests/Features/Urls.cs
@@ -1,4 +1,4 @@
-﻿using SFA.DAS.ApprenticeCommitments.Web.Pages.IdentityHashing;
+﻿using SFA.DAS.ApprenticeCommitments.Web.Identity;
 
 namespace SFA.DAS.ApprenticeCommitments.Web.UnitTests.Features
 {
