+++ src/SFA.DAS.ApprenticeCommitments.Web/Pages/Apprenticeships/CannotConfirm.cshtml.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using SFA.DAS.ApprenticeCommitments.Web.Pages.IdentityHashing;
+using SFA.DAS.ApprenticeCommitments.Web.Identity;
 using SFA.DAS.ApprenticeCommitments.Web.Services;
 
 namespace SFA.DAS.ApprenticeCommitments.Web.Pages.Apprenticeships
